refactor(number-theory): isolate modular residue contracts
